network: calculate network size wrt seeds
Problem:

In case when node is disconnected from the network (and all of its peers are
blacklisted due to consequtive failed connection attempts), estimated
network size is 1 (the node itself):
```
2025-05-21 02:36:45.182	info	network/server.go:569	proto loop	{"loopCnt": 53, "netSize": 1, "optimalN": -4294967296, "peerN": 0, "minPeers": 2, "maxPeers": 9, "peerT": 1000}
```

The problem is that we now for sure that seed nodes should be a part of
the network, and further calculations that rely on (*DefaultDiscovery).networkSize
should take it into account.

Solution:

Take into account seed nodes while calculating estimated network size.

Discovered during debugging of #3902.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>