native: don't process seek results after IC finalization
Once interop context is finilized, the results of storage iterators are
not relevant and there's no sence to send them to a processing channel.

The problem of #3927 is:
    1. TestNEO_GetCandidates calls `getAllCandidates` method of native
       NeoToken and iterates through the resulting interop via tx script.
       The main point is that it does not read all items from the iterator.
    2. Transaction is being persisted. Starting from #3919, native cache is
       not locked during private chain cache persist.
    3. Some of SEEK results still remain in the storage and are pushed to the
       iterator processing routine even during/after DAO persist.

The provblem is that `(*NEO).getAllCandidatesCall` uses
`(*Policy).IsBlocked` to handle iterator results, whereas `IsBlocked`
performs an access to the underlying native cache. At the same time
persisting routine clears this cache during persist.

The solution is to properly wait for interop context finalization before
storeBlock's cache persist.

Close #3927.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>