*: fix/check JSON stitching for complex types
Discovered when changing omitempty to omitzero which works differently for
util.Uint256 (0x00000... vs nothing). In this case empty tx metadata can
JSON into '{}' and we end up with '{, ... }' in JSON for the overall
structure. This can't happen in other cases, all structures there have
something that is always JSONized, but for tx meta we do want to omit
blockhash when it's missing (which can be considered as deviation from C#
before the patch). Empty metadata is relevant for not-yet-onchain
transactions.

Signed-off-by: Roman Khimov <roman@nspcc.ru>