vm: make loadbse64/loadhex work with files
Large scripts (10K+ instructions) can't be pasted to the VM CLI
properly on Linux due to some bug in urfave/cli, the result is that the
terminal screen is continiously updating with the large amount of data.

Add an ability to loadbase64 and loadhex commands to read argument from
a file.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>