context: reverse the order of parameters
An attempt to use `neo-go util sendtx` for a transaction finished (that is
having Parameters filled with signatures because there are enough signatures
to do that) the C# code ends up with

    Invalid signature (-508) - witness #0: signature check failed: invalid signature

The reason is that C# node uses reverse parameter order and order matter a lot
for signatures (since keys in multisignature contract are ordered as well).

This fix technically breaks compatibility with the old code, but this shouldn't
be a problem since this command is only used for offline signing which is a
rare case, most of the time `wallet sign` sends the transaction immediately.
In any event using consistent versions of NeoGo solves it.

Signed-off-by: Roman Khimov <roman@nspcc.ru>