audit/neofs: fix cursor reset on temporary search failures
Consider an error from search like this:

    Retry 0: failed to search objects: connection: no healthy client

it kills the `cursor` by making it an empty string. This effectively resets the search,
but block counter is still ticking, so this leads to incorrect indexes like

    search returned 1000 results, previous height 10185999

in a container that has the maximum index stored of 8559864.

Same thing can happen in neofs package, so fixed there as well.

Signed-off-by: Roman Khimov <roman@nspcc.ru>